feat: add polkadot-js as a dependency
