+++ .gitignore
@@ -1,3 +1,3 @@
-/dist
-/node_modules
+dist
+node_modules
 package-lock.json
