+++ package.json
@@ -11,6 +11,7 @@
   },
   "author": "Ideal Labs <hello@idealabs.network>",
   "dependencies": {
+    "@polkadot/api": "^13.2.1",
     "axios": "^1.7.7"
   },
   "devDependencies": {
